// src/app/page/master/suku/suku.interface.ts
import { EventEmitter } from '@angular/core';
export interface Suku {
  kdDepartemen?: string;
  kdStatusHead?: 0;
  kode?: 0;
  kodeExternal?: string;
  namaExternal?: string;
  namaStatus?: string;
  reportDisplay?: string;
  statusEnabled?: true;
  version?: 0
  noRec?: string;
  label?: any;
  value?: any;
}